@@ -15,10 +15,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+13
+version: 1.0.0+14
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,39 +27,43 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
+  #sticky_grouped_list: ^1.3.0
+  #grouped_list: ^4.0.0
+  #flutter_local_notifications: ^5.0.0
+  #badges: ^1.1.6
+  flutter_geofire: ^2.0.0
   cupertino_icons: ^1.0.1
-  sticky_grouped_list: ^1.3.0
-  grouped_list: ^4.0.0
-  google_maps_flutter: ^2.0.2
-  firebase_core: ^1.0.1
-  firebase_auth: ^1.0.1
-  google_sign_in: ^4.5.9
-  geolocator: ^5.1.1+1
-  location: ^4.1.1
-  cloud_firestore: ^2.0.0
-  fluttertoast: ^7.1.6
-  ndialog: ^3.2.1
+  google_maps_flutter: ^2.0.6
+  firebase_core: ^1.4.0
+  firebase_auth: ^3.0.1
+  google_sign_in: ^5.0.4
+  geolocator: ^7.3.0
+  geocoding: ^2.0.0
+  location: ^4.3.0
+  cloud_firestore: ^2.3.0
+  fluttertoast: ^8.0.7
   url_launcher: ^6.0.0
   http: ^0.13.1
-  provider: ^4.3.3
-  flutter_local_notifications: ^5.0.0
+  provider: ^5.0.0
   shared_preferences: ^2.0.3
-  cpfcnpj: ^1.0.3
-  image_picker: ^0.7.0
-  firebase_storage: ^8.0.0
-  badges: ^1.1.6
-  font_awesome_flutter: ^8.11.0
-  permission_handler: ^6.1.3
-  shortid: ^0.1.1
-  flutter_geofire: ^2.0.0
+  flux_validator_dart: ^2.0.0
+  image_picker: ^0.8.2
+  firebase_storage: ^9.0.0
+  font_awesome_flutter: ^9.0.0
+  permission_handler: ^8.1.4
+  short_uuids: ^2.0.0
   firebase_messaging: ^10.0.0
-  assets_audio_player: ^3.0.2
+  firebase_analytics: ^8.1.2
+  assets_audio_player: ^3.0.3+6
   intl: ^0.17.0
-  smooth_star_rating: 1.1.1
-  firebase_database: ^6.0.0
-  maps_toolkit: ^1.1.0+2
-  firebase_crashlytics: ^2.0.1
-  wakelock: ^0.2.1+1
+  flutter_rating_bar: ^4.0.0
+  firebase_database: ^7.1.2
+  #calcula a area
+  maps_toolkit: ^2.0.0
+  firebase_crashlytics: ^2.1.1
+  wakelock: ^0.5.3+3
+  flutter_local_notifications: ^5.0.0+4
+  paginate_firestore: ^1.0.0
   flutter_mobx:
   mobx:
 
