@@ -5,12 +5,8 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:fogaca_app/Model/Motoboy.dart';
 import 'package:fogaca_app/Notificacao/PushNotificacao.dart';
-import 'package:fogaca_app/Store/StoreDadosUsuario.dart';
-import 'package:fogaca_app/Widget/Toast.dart';
 import 'package:mobx/mobx.dart';
 import 'package:provider/provider.dart';
-import 'package:smooth_star_rating/smooth_star_rating.dart';
-
 
 class RatingTabPage extends StatefulWidget {
 
@@ -23,8 +19,8 @@ class _RatingTabPageState extends State<RatingTabPage> with AutomaticKeepAliveCl
 
   PushNotificacao pushNotificacao= PushNotificacao();
   Motoboy motoboyLogado=Motoboy();
-  String estrela;
-  double estrelaDouble;
+  String? estrela;
+  double ?estrelaDouble;
 
 
 
@@ -36,8 +32,8 @@ class _RatingTabPageState extends State<RatingTabPage> with AutomaticKeepAliveCl
       if(motoboyLogado.estrela!=null){
         estrelaDouble = double.parse(motoboyLogado.estrela);
       }
-      double mod = pow(10.0, 2);
-      estrela = ((estrelaDouble * mod).round().toDouble() / mod).toString();
+      num mod = pow(10.0, 2);
+      estrela = ((estrelaDouble! * mod).round().toDouble() / mod).toString();
     }
 
 
@@ -47,22 +43,8 @@ class _RatingTabPageState extends State<RatingTabPage> with AutomaticKeepAliveCl
           children: [
 
             SizedBox(height: 200.0,),
-           Center(
-            child: Text(estrela!=null?estrela:"",style:new TextStyle(fontSize: 65.0),),
+         //  Center(child: Text(estrela!=null?estrela:"",style:new TextStyle(fontSize: 65.0),),),
 
-          ),
-            SmoothStarRating(
-              rating:estrelaDouble!=null?estrelaDouble:5.0,
-              borderColor: Colors.red[800],
-              color: Colors.red[800],
-              isReadOnly: true,
-              allowHalfRating: false,
-              starCount: 5,
-              size: 50,
-              onRated: (value)
-              {
-              },
-            ),
         ]),
 
       ),
