@@ -2,21 +2,23 @@ import 'package:flutter/material.dart';
 
 class WIBusy extends StatelessWidget{
   bool busy =false;
-  Widget child;
+  Widget ?child;
 
   WIBusy({
-    @required this.busy,
-    @required this.child,
+    required this.busy,
+          this.child,
   });
   @override
   Widget build(BuildContext context) {
 
-    return busy
-        ? Container(
+    return busy?
+    Container(
+      width: double.infinity,
+      height: 100,
       child:Center(
         child: CircularProgressIndicator(),
       ),
-    ):child;
+    ):child!;
 
   }
 }
