@@ -2,15 +2,18 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
+import 'package:firebase_analytics/firebase_analytics.dart';
+import 'package:firebase_analytics/observer.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter/src/foundation/assertions.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:fogaca_app/Notificacao/PushNotificacao.dart';
-import 'package:fogaca_app/Widget/WIListPorData.dart';
+
 import 'Pages_pedido/Pedidos_em_Entrega.dart';
 import 'Pages_user/Tela_Login.dart';
 import 'Pages_user/Tela_Cad_Moto.dart';
@@ -41,8 +44,8 @@ Future<void> _firebaseMessagingBackgroundHandler(RemoteMessage message) async {
   });
 }
 
-AndroidNotificationChannel channel;
-FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin;
+AndroidNotificationChannel ?channel;
+FlutterLocalNotificationsPlugin? flutterLocalNotificationsPlugin;
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -61,16 +64,16 @@ void main() async {
   await Firebase.initializeApp();
   final DadosFirestore=Dados_usuario();
   await FirebaseCrashlytics.instance.setCrashlyticsCollectionEnabled(true);
-  Function originalOnError = FlutterError.onError;
+  FlutterExceptionHandler? originalOnError = FlutterError.onError;
   FlutterError.onError = (FlutterErrorDetails errorDetails) async {
     await FirebaseCrashlytics.instance.recordFlutterError(errorDetails);
-    originalOnError(errorDetails);
+    originalOnError!(errorDetails);
   };
 
-  await flutterLocalNotificationsPlugin
+  await flutterLocalNotificationsPlugin!
       .resolvePlatformSpecificImplementation<
           AndroidFlutterLocalNotificationsPlugin>()
-      ?.createNotificationChannel(channel);
+      ?.createNotificationChannel(channel!);
 
   /// Update the iOS foreground notification presentation options to allow
   /// heads up notifications.
@@ -82,79 +85,42 @@ void main() async {
 
   runApp(
     MultiProvider(providers: [
-      Provider<StoreDadosUsuario>(
-        create: (_) => StoreDadosUsuario(),
-      ),
+      // Provider<StoreDadosUsuario>(
+      //   create: (_) => StoreDadosUsuario(),
+      // ),
       ChangeNotifierProvider<ThemeChanger>(
         create: (_) => ThemeChanger(),
       ),
       ChangeNotifierProvider<Dados_usuario>(
         create: (_) => Dados_usuario(),
       ),
-      StreamProvider(create: (context)=>DadosFirestore.GetClientes()),
+     // StreamProvider(create: (context)=>DadosFirestore.GetClientes()),
     ], child: MyApp()),
   );
 }
 
 class MyApp extends StatelessWidget {
-  PushNotificacao pushNotificacao = PushNotificacao();
   @override
   Widget build(BuildContext context) {
-    pushNotificacao.initialize(context);
-    // TODO: implement build
-    bool darkThemeEnabled = Provider.of<ThemeChanger>(context).isDark();
+    FirebaseAnalytics analytics = FirebaseAnalytics();
     return MaterialApp(
-      //Thema LIGHT
       theme: ThemeData(
-          dialogBackgroundColor: Colors.white,
-          brightness: Brightness.light,
-          primarySwatch: Colors.red,
-          accentColor: Colors.redAccent,
-          cardColor: Colors.white,
-          fontFamily: "Brand Bold",
-          textTheme: TextTheme(
-            bodyText1: TextStyle(),
-            bodyText2: TextStyle(),
-          ).apply(
-            bodyColor: Colors.black54,
-            displayColor: Colors.white60,
-          ),
-          visualDensity: VisualDensity.adaptivePlatformDensity),
-      //Thema Dark
-      darkTheme: ThemeData(
-        brightness: Brightness.dark,
-        dialogBackgroundColor: Colors.grey[800],
-        fontFamily: "Brand Bold",
-        cardColor: Colors.black87,
-        accentColor: Colors.redAccent,
-        primarySwatch: createMaterialColor(Color(4288088072)),
-        textTheme: TextTheme(
-          bodyText1: TextStyle(),
-          subtitle1: TextStyle(color: Colors.black87),
-        ).apply(
-          bodyColor: Colors.white,
-        ),
-        // accentColor:Colors.redAccent,
-        visualDensity: VisualDensity.adaptivePlatformDensity,
+        primarySwatch:createMaterialColor(Color(4288088072)),
+        scaffoldBackgroundColor: Color(0xFF403939),
+        cardColor: Color(0xFF403939),
+        canvasColor: Color(0xFF403939),
+         // visualDensity: VisualDensity.adaptivePlatformDensity
       ),
-      themeMode:ThemeMode.dark,
       debugShowCheckedModeBanner: false,
       title: "Fogaça Motoboys",
-      //verificando se o usuario esta logado
-      initialRoute: FirebaseAuth.instance.currentUser == null
-          ? Tela_Login.idScreen
-          : SplashScreen.idScreen,
+      navigatorObservers: [FirebaseAnalyticsObserver(analytics: analytics)],
+     home:SplashScreen() ,
       routes: {
         SplashScreen.idScreen: (context) => SplashScreen(),
         Tela_Cadastro.idScreen: (context) => Tela_Cadastro(),
         Tela_Login.idScreen: (context) => Tela_Login(),
         Mapa_Home.idScreen: (context) => Mapa_Home(),
         Pedidos_em_Entrega.idScreen: (context) => Pedidos_em_Entrega(),
-
-        ///  MotoInfo.idScreen:(context)=>MotoInfo(),
-        // Pedidos_em_Andamento.idScreen:(context)=>Pedidos_em_Andamento(),
-        //Tela_Cad_Loc_Loja.idScreen:(context)=>Tela_Cad_Loc_Loja(),
-        //Tela_Pesquisa.idScreen:(context)=>Tela_Pesquisa(),
       },
     );
   }
@@ -162,7 +128,7 @@ class MyApp extends StatelessWidget {
   //Color
   MaterialColor createMaterialColor(Color color) {
     List strengths = <double>[.05];
-    Map swatch = <int, Color>{};
+    Map<int, Color> swatch = <int, Color>{};
     final int r = color.red, g = color.green, b = color.blue;
 
     for (int i = 1; i < 10; i++) {
