@@ -6,12 +6,12 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
 class CPButtonSFundo extends StatelessWidget{
 
-  final String text;
-  final String  imagem;
-  final Function() callback;
-  final double width;
-  final double height;
-  final IconData icon;
+  final String ?text;
+  final String  ?imagem;
+  final Function() ?callback;
+  final double ?width;
+  final double ?height;
+  final IconData ?icon;
   CPButtonSFundo({
     @required this.text,
     this.callback,
@@ -30,9 +30,9 @@ class CPButtonSFundo extends StatelessWidget{
           Container(
             height: height,
             width: width,
-            child:Image.asset(imagem),
+            child:Image.asset(imagem!),
           ),
-          Text(text)
+          Text(text!)
         ],
       ),
       onPressed:callback
