@@ -6,12 +6,12 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
 class CPButtonText extends StatelessWidget{
 
-  final String text;
-  final String  imagem;
-  final Function() callback;
-  final double width;
-  final double height;
-  final IconData icon;
+  final String ?text;
+  final String  ?imagem;
+  final Function() ?callback;
+  final double ?width;
+  final double ?height;
+  final IconData ?icon;
   CPButtonText({
     @required this.text,
     this.callback,
@@ -25,7 +25,7 @@ class CPButtonText extends StatelessWidget{
     // TODO: implement build
     return  TextButton(
         child:Text(
-          text,
+          text!,
               style:TextStyle(
                   color:  Color(0xFFC83535),
                 fontFamily: "Brand-Regular",
