@@ -1,9 +1,9 @@
 
 class PlacePredictions
 {
-  String secondary_text;
-  String main_text;
-  String place_id;
+  var secondary_text;
+  var main_text;
+  var place_id;
 
   PlacePredictions({this.secondary_text,this.main_text,this.place_id});
 
